refactor(frontend): reorganize JS files into canonical structure (issue #147) (#148)
* Initial plan

* refactor: reorganize frontend files per issue #147

---------
