--- frontend/assets/js/components/elements/controllers/HeaderController.jsx
@@ -1,4 +1,0 @@
-/** Re-exports {@link HeaderController} for use as a `.jsx` module. */
-import HeaderController from './HeaderController.js';
-
-export default HeaderController;
