--- frontend/assets/js/components/elements/helpers/HeaderHelper.js
@@ -1,4 +1,0 @@
-/** Re-exports {@link HeaderHelper} for use as a `.js` module. */
-import HeaderHelper from './HeaderHelper.jsx';
-
-export default HeaderHelper;
