+++ frontend/assets/js/utils/HashRouteResolver.js
